[1.x] fix: Parse `remote_connect_timeout_secs` as a double (#2951)
* fix: Parse `remote_connect_timeout_secs` as a double


* test: Test parsing of `.condarc`


* lint: Format with black


---------

+++ libmamba/src/core/curl.cpp
@@ -22,7 +22,7 @@ namespace mamba
             CURL* handle,
             const std::string& url,
             const bool set_low_speed_opt,
-            const long connect_timeout_secs,
+            const double connect_timeout_secs,
             const bool set_ssl_no_revoke,
             const std::optional<std::string>& proxy,
             const std::string& ssl_verify
@@ -122,7 +122,7 @@ namespace mamba
         bool check_resource_exists(
             const std::string& url,
             const bool set_low_speed_opt,
-            const long connect_timeout_secs,
+            const double connect_timeout_secs,
             const bool set_ssl_no_revoke,
             const std::optional<std::string>& proxy,
             const std::string& ssl_verify
@@ -391,7 +391,7 @@ namespace mamba
     void CURLHandle::configure_handle(
         const std::string& url,
         const bool set_low_speed_opt,
-        const long connect_timeout_secs,
+        const double connect_timeout_secs,
         const bool set_ssl_no_revoke,
         const std::optional<std::string>& proxy,
         const std::string& ssl_verify
